--- src/main/java/mops/foren/domainmodel/Forum.java
@@ -1,4 +1,0 @@
-package mops.foren.domainmodel;
-
-public class Forum {
-}
