--- src/main/java/mops/foren/domainmodel/Modus.java
@@ -1,4 +1,0 @@
-package mops.foren.domainmodel;
-
-public class Modus {
-}
