--- src/main/java/mops/foren/domainmodel/Permission.java
@@ -1,4 +1,0 @@
-package mops.foren.domainmodel;
-
-public enum Permission {
-}
