--- src/main/java/mops/foren/domainmodel/Topics.java
@@ -1,4 +1,0 @@
-package mops.foren.domainmodel;
-
-public class Topics {
-}
