+++ src/main/java/mops/foren/model/Course.java
@@ -1,0 +1,4 @@
+package mops.foren.model;
+
+public class Course {
+}
