+++ src/main/java/mops/foren/model/Forum.java
@@ -1,0 +1,4 @@
+package mops.foren.model;
+
+public class Forum {
+}
