+++ src/main/java/mops/foren/model/Modus.java
@@ -1,0 +1,4 @@
+package mops.foren.model;
+
+public class Modus {
+}
