+++ src/main/java/mops/foren/model/Permission.java
@@ -1,0 +1,4 @@
+package mops.foren.model;
+
+public enum Permission {
+}
