+++ src/main/java/mops/foren/model/Thread.java
@@ -1,0 +1,4 @@
+package mops.foren.model;
+
+public class Thread {
+}
