+++ src/main/java/mops/foren/services/DomainService.java
@@ -1,4 +1,4 @@
-package mops.foren.domainmodel;
+package mops.foren.services;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
