+++ src/main/java/mops/foren/services/TopicService.java
@@ -1,4 +1,4 @@
-package mops.foren.domainmodel;
+package mops.foren.services;
 
 @DomainService
 public class TopicService {
